parser: Actually use the SIMD
The classify_ssse3 and classify_avx2 were built with the default
target's configuration, so in our case x86_64 with sse2. So some of the
intrinsics were using fallback code instead of the actual SIMD instructions.
The ssse3 variant is mostly sse2, so it was still faster than fallback,
but avx2 was slower...

Now we are finally seeing better speeds.