parser: Handle buffer end after <!
If we get the exception because of "<!", which is start of comment or
CDATA, but the buffer ends right there and we can not determine whether
it is comment or CDATA, we would return parsing error, which was wrong.
Now we will stay in the exception state until we have more data.